improve animation player
